# docs/swagger.yaml
basePath: /
definitions:
...
  merchant.AppointmentStatsResponse:
    properties:
      month:
        description: 本月预约数
        example: 450
        type: integer
      status:
        additionalProperties:
          type: integer
        description: 各状态预约数
        type: object
      today:
        description: 今日预约数
        example: 15
        type: integer
      week:
        description: 本周预约数
        example: 120
        type: integer
    type: object
...
  merchant.CreateStaffRequest:
    properties:
      avatar:
        type: string
      description:
        type: string
      name:
        type: string
      position:
        type: string
      specialties:
        type: string
    required:
    - name
    type: object
  merchant.DailyRevenue:
    properties:
      amount:
        description: 当日收入
        example: 1200.5
        type: number
      date:
        description: 日期
        example: "2023-06-15"
        type: string
    type: object
...
  merchant.RevenueStatsResponse:
    properties:
      daily_revenue:
        description: 每日收入数据
        items:
          $ref: '#/definitions/merchant.DailyRevenue'
        type: array
      service_revenue:
        description: 服务收入分布
        items:
          $ref: '#/definitions/merchant.ServiceRevenue'
        type: array
      total_revenue:
        description: 总收入
        example: 15000.5
        type: number
    type: object
  merchant.ServiceRevenue:
    properties:
      amount:
        description: 该服务总收入
        example: 5000
        type: number
      service_id:
        description: 服务ID
        example: 1
        type: integer
      service_name:
        description: 服务名称
        example: 基础护理
        type: string
    type: object
...
  merchant.UpdateStaffRequest:
    properties:
      avatar:
        type: string
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      position:
        type: string
      specialties:
        type: string
    type: object
...
  models.Staff:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      merchantID:
        type: integer
      name:
        type: string
      position:
        type: string
      specialties:
        type: string
      updatedAt:
        type: string
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 预约系统API文档
  title: Appointment System API
  version: "1.0"
paths:
...
  /api/merchant/staff:
    get:
      consumes:
      - application/json
      description: 获取当前商家的所有员工列表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 员工列表
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Staff'
                  type: array
              type: object
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 获取员工列表失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取员工列表
      tags:
      - 商家员工管理
    post:
      consumes:
      - application/json
      description: 为当前商家创建新员工
      parameters:
      - description: 员工信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/merchant.CreateStaffRequest'
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 员工创建成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Staff'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 创建员工失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建员工
      tags:
      - 商家员工管理
  /api/merchant/staff/{staffId}:
    delete:
      consumes:
      - application/json
      description: 删除员工（软删除或硬删除）
      parameters:
      - description: 员工ID
        example: 123
        in: path
        name: staffId
        required: true
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 员工删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 无效的员工ID
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 员工不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 删除员工失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除员工
      tags:
      - 商家员工管理
    put:
      consumes:
      - application/json
      description: 更新员工信息
      parameters:
      - description: 员工ID
        example: 123
        in: path
        name: staffId
        required: true
        type: integer
      - description: 员工更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/merchant.UpdateStaffRequest'
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Staff'
              type: object
        "400":
          description: 无效的员工ID | 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 员工不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 更新员工失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新员工
      tags:
      - 商家员工管理
  /api/merchant/stats/appointments:
    get:
      description: 获取当前商户的预约统计数据，包括今日、本周、本月预约数以及各状态预约数
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: '开始日期 (格式: YYYY-MM-DD)'
        example: '"2023-06-01"'
        in: query
        name: start_date
        type: string
      - description: '结束日期 (格式: YYYY-MM-DD)'
        example: '"2023-06-30"'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回预约统计数据
          schema:
            $ref: '#/definitions/merchant.AppointmentStatsResponse'
        "500":
          description: 获取数据失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取预约统计数据
      tags:
      - 商户-数据统计
  /api/merchant/stats/revenue:
    get:
      description: 获取当前商户的营收统计数据，包括总收入、每日收入趋势和各服务收入分布
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: '开始日期 (格式: YYYY-MM-DD)'
        example: '"2023-06-01"'
        in: query
        name: start
        required: true
        type: string
      - description: '结束日期 (格式: YYYY-MM-DD)'
        example: '"2023-06-30"'
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回营收统计数据
          schema:
            $ref: '#/definitions/merchant.RevenueStatsResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 获取数据失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取营收统计数据
      tags:
      - 商户-数据统计
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"